+++ packages/nexus/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @plitzi/nexus
 
+## 0.32.1
+
+### Patch Changes
+
+- v0.32.1
+
 ## 0.32.0
 
 ### Minor Changes