+++ packages/sdk-interactions/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @plitzi/sdk-interactions
 
+## 0.32.1
+
+### Patch Changes
+
+- v0.32.1
+- Updated dependencies
+  - @plitzi/sdk-auth@0.32.1
+  - @plitzi/sdk-dev-tools@0.32.1
+  - @plitzi/sdk-event-bridge@0.32.1
+  - @plitzi/sdk-shared@0.32.1
+
 ## 0.32.0
 
 ### Minor Changes